% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr.R
\name{calculate_clr}
\alias{calculate_clr}
\title{Calculate CLR}
\usage{
calculate_clr(mat, base = "e", denom_method = "geomean")
}
\arguments{
\item{mat}{an D x M matrix of D target IDs and
M samples}

\item{base}{what should the base of the logarithm be?
currently only supports base "e" and base 2.}

\item{denom_method}{either 'geomean' or 'DESeq2' to
use either the geometric mean of all features as the
denominator, or the DESeq2-style size factors as the denominator
(equivalent to using standard DESeq2-style normalization, 
which is used in standard sleuth)}
}
\value{
D x M matrix of CLR-transformed values,
  centered on the geometric mean of all features
  within each sample
}
\description{
This function calculates the centered logratio
transformation (subtracting geometric mean of all
features within each sample).
}
