% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{deseq_size_factors}
\alias{deseq_size_factors}
\title{Get DESeq2 Size Factors}
\usage{
deseq_size_factors(x, denoms = NULL)
}
\arguments{
\item{x}{a D x M matrix with D features (the denominator features) and
M samples}

\item{denoms}{a vector of either indices or feature names that match row names
of the matrix. These are the ones that will be used for calculating the
compositional normalization using the DESeq2-method.}
}
\value{
a vector of the size factors for each sample
}
\description{
Function to return the DESeq2-style size factors of a collection of features.
The geometric mean is calculated for feature across samples. Then the
ratio of every sample observation to the geometric mean across samples is
calculated. From there, the median of ratios within each sample is selected
as the size factor.
}
\references{
For a discussion of why the DESeq2 size factor is compatible
  with Compositional Normalization, see
  \url{https://dx.doi.org/10.1093/bioinformatics/bty175}

For the original DESeq2 method, see
  \link[DESeq2]{estimateSizeFactorsForMatrix}
}
