% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqlr.R
\name{find_iqlr_denoms}
\alias{find_iqlr_denoms}
\title{Find IQLR Denominators}
\usage{
find_iqlr_denoms(mat, base = "e")
}
\arguments{
\item{mat}{a D x M matrix, with D features and M samples}

\item{base}{either "e" (for natural log scale) or "2" for base-2 logarithm}
}
\value{
a numeric vector containing the indices of the features that
  form the set to be used as the denominator for the IQLR transformation
}
\description{
This function identifies which features in a matrix 
are suitable set to use as a denominator for the
'Interquartile Logratio' (IQLR) transformation.
These are the features that have variance in the
interquartile range after transforming the data using
the centered logratio (CLR) transformation.
This was originally introduced in ALDEx2 in its paper.
}
\references{
For the ALDEx2 paper discusing the IQLR transformation, see
  \url{https://dx.doi.org/10.1186\%2F2049-2618-2-15}
}
