% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{geomean}
\alias{geomean}
\title{Geometric mean}
\usage{
geomean(x)
}
\arguments{
\item{x}{numeric vector of positive numbers.}
}
\value{
the geometric mean of x
}
\description{
Function to return the geometric mean. By default,
it only calculates on positive numbers, and gives a
warning if there are any NA values or any numbers
that are zero or negative.
}
