% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_denom_names}
\alias{get_denom_names}
\title{Get Denom Name(s)}
\usage{
get_denom_names(obj)
}
\arguments{
\item{obj}{the sleuth object}
}
\value{
The name of the denominator(s) for the sleuth object
}
\description{
Function to retrieve the denominator(s) for a sleuth object
processed using compositional data analysis. If the analysis
was done using CLR (centered logratio), the denominator name is
all.
}
