% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{get_lr_function}
\alias{get_lr_function}
\title{Get Logration Function}
\usage{
get_lr_function(type = "alr", denom_name = NULL)
}
\arguments{
\item{type}{either "alr" or "clr" (also accepts "ALR" / "CLR")}

\item{denom_name}{a character vector of target IDs that
will be the denominator. More than one is acceptable, and
you can also input a numeric vector of row numbers instead.}
}
\value{
a transformation function ready to be used for the
 'transform_fun' option in 'sleuth_prep'
}
\description{
This helps you create a function that can be input as a
transformation function for 'sleuth_prep'.
}
