% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeros.R
\name{impute_zeros}
\alias{impute_zeros}
\title{Impute Rounded Zeros}
\usage{
impute_zeros(mat, method = "multiplicative", delta = NULL,
  impute_proportion = 0.65)
}
\arguments{
\item{mat}{a numeric matrix to be imputed, with N
targets and M samples. This method assumes that N > M.
N x M and M x N matrices are both valid inputs.}

\item{method}{the choice of how to impute the rounded zeros.
only "multiplicative" and "additive" is supported at this time.}

\item{delta}{the value to impute; if NULL,
delta = impute_proportion x (the minimum value in a sample)}

\item{impute_proportion}{the proportion of the minimum value
that should be the delta value (default = 65\%).}
}
\value{
a matrix with the same dimensions as input and all zero
  values imputed by sample.
}
\description{
This uses various strategies to impute zeros
that are considered "rounded zeros", that is, values
that are zero because the true value is positive but below
the detection limit (i.e. the sequencing depth). See the
references for more information.
}
