% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqlr.R
\name{iqlr_transformation}
\alias{iqlr_transformation}
\title{Calculate IQLR}
\usage{
iqlr_transformation(mat, base = "e", remove_zeros = FALSE, delta = NULL,
  impute_proportion = 0.65)
}
\arguments{
\item{mat}{an D x M matrix of D target IDs and
M samples}

\item{base}{what should the base of the logarithm be?
currently only supports base "e" and base 2.}

\item{delta}{a number that is the imputed value. If \code{NULL},
delta = impute_proportion * (minimum value in sample)}

\item{impute_proportion}{percentage of minimum value that
becomes the imputed value. Only used if delta is \code{NULL}}
}
\value{
(D - n - z) x M matrix of ALR-transformed
  values, with n equal to the number of denominator values
  and z are the number of rows with essential zeros.
}
\description{
This function calculates the interquartile logratio
transformation. This function first processes the data
to handle zeros. It then calculates which features
have the variance across samples within the interquartile range
after the CLR transformation. It then uses calculate_alr
internally after processing the data to handle
zeros. Essential zeros (i.e. target IDs with 
zeros in all samples) are excluded, and
round zeros (i.e. target IDs with at least one
zero value) are imputed using the mulitplicative
method.
}
