% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sleuth.R
\name{make_lr_sleuth_object}
\alias{make_lr_sleuth_object}
\title{Make Compositional Sleuth Object}
\usage{
make_lr_sleuth_object(sample_to_covariates, denom_name, lr_type = "alr",
  full_model = NULL, beta = NULL, null_model = NULL,
  run_models = FALSE, ...)
}
\arguments{
\item{sample_to_covariates, }{the sample_to_covariates \code{data.frame} for
sleuth}

\item{denom_name, }{target ID names or index numbers of features to be used
for the denominator when using compositional analysis; this argument is
required for the ALR transformation. Using 'clr' or 'iqlr' for 'lr_type' 
overrides this argument (and the CLR / IQLR transformation will be used).
If 'best' is used, then the internal function \code{choose_best_denoms}
will be used to identify the feature(s) with the most consistent abundance
across all samples in the experiment. See \link{choose_denom} for extra
options.}

\item{lr_type, }{either "alr", "clr", or "iqlr" ("ALR" / "CLR" / "IQLR" are
also accepted), indicating additive, centered, or interquartile logratio
transformation}

\item{full_model, }{the full model for sleuth}

\item{beta, }{the beta you wish to use for the Wald test. If \code{NULL}
(the default), the Wald test will be skipped.}

\item{null_model, }{the null model to be the baseline for the LR test.
If \code{NULL} (the default), this step is skipped.}

\item{run_models}{boolean to see if the modeling step should be done.
If \code{FALSE}, the default, only sleuth_prep is done.}

\item{...}{extra options that will tweak the analysis, specifically for
\code{\link{get_lr_functions}}, \code{sleuth_prep}, and \code{sleuth_fit}.
for details on which options can be specified for \code{sleuth_prep} and
\code{sleuth_fit}, please see ?get_lr_functions, ?sleuth::sleuth_prep or
?sleuth::sleuth_fit for details. Note that for \code{sleuth_prep},
\code{read_bootstrap_tpm} and \code{extra_bootstrap_summary} are \code{TRUE}
by default to allow for modeling estimated or TPMs downstream.}
}
\value{
a sleuth object that has been prepped using the compositional
  analysis for the normalization and transformation steps, and fitted
  using the full model (if \code{run_models} is \code{TRUE}) and null model
  (if specified). It will also run the Wald test (if \code{beta} is
  specified) and the LR test (if applicable).
}
\description{
This is a wrapper function for the sleuth pipeline that applies
the compositional normalization approach. Many of the arguments
are ones that will be input into different parts of the sleuth
pipeline: \code{sleuth_prep}, \code{sleuth_fit}, \code{sleuth_wt},
and \code{sleuth_lrt}.
}
