% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{norm_identity}
\alias{norm_identity}
\title{Norm Identity}
\usage{
norm_identity(mat)
}
\description{
Function to return size factors of just 1
for sleuth normalization (i.e. lack of
normalization). Normalization is not
necessary for logratio transformations.
}
