% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_extra_opts.R
\name{process_extra_opts}
\alias{process_extra_opts}
\title{Process extra options for main sleuth-ALR function}
\usage{
process_extra_opts(...)
}
\arguments{
\item{...}{extra arguments passed from \code{\link{make_lr_sleuth_object}}}
}
\description{
This internal function processes all of the extra arguments for various
steps of the sleuth-ALR pipeline.
}
\details{
Four sets of arguments will be processed: sleuth_prep (prep_opts),
  sleuth_fit (fit_opts), get_lr_functions (alr_opts), and choose_denom
  (choose_opts). See the documentation for those functions if you want a
  list of arguments to use and how they are used.
}
