% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeros.R
\name{remove_essential_zeros}
\alias{remove_essential_zeros}
\title{Remove Essential Zeros}
\usage{
remove_essential_zeros(mat)
}
\arguments{
\item{mat}{an N x M numeric matrix.}
}
\value{
an (N-z) x M numeric matrix, with z
  equal to the number of rows with all zero values.
}
\description{
This function excludes "essential zeros", that is,
target IDs that are assumed to be silent and
non-expressed.
}
