% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sleuth.R
\name{sleuth_alr_results}
\alias{sleuth_alr_results}
\title{Return sleuth-ALR Results Table}
\usage{
sleuth_alr_results(obj, test, test_type = "wt", which_model = "full",
  show_all = TRUE, pval_aggregate = obj$pval_aggregate,
  weight_func = "best")
}
\arguments{
\item{obj}{a \code{sleuth} object}

\item{test}{a character string denoting the test to extract. Possible tests can be found by using \code{models(obj)}.}

\item{test_type}{'wt' for Wald test or 'lrt' for Likelihood Ratio test.}

\item{which_model}{a character string denoting the model. If extracting a wald test, use the model name.
Not used if extracting a likelihood ratio test.}

\item{show_all}{if \code{TRUE} will show all transcripts (not only the ones
passing filters). The transcripts that do not pass filters will have
\code{NA} values in most columns.}

\item{pval_aggregate}{if \code{TRUE} and both \code{target_mapping} and \code{aggregation_column} were provided,
to \code{sleuth_prep}, use lancaster's method to aggregate p-values by the \code{aggregation_column}.}

\item{weight_func}{if \code{pval_aggregate} is \code{TRUE}, then this is used to weight the p-values for
lancaster's method. This must be either the string 'best' or it must be a function that takes the
observed means of the transcripts as the only defined argument.}
}
\value{
a \code{data.frame} with the same specification as found in \code{sleuth_results}. See \code{sleuth_results} for
  details.
}
\description{
This is a wrapper function for \code{sleuth_results} that properly
handles the unique features of sleuth-ALR-transformed data.
}
\details{
For the transcript-level analysis, this produces the same results as the default \code{sleuth_results}.
  However, using the default function for p-value aggregation is incompatible with the standard sleuth-ALR
  transformation. Sleuth-ALR logratios typically include negative values (any feature that is less abundant than the
  chosen 'reference feature(s)' will yield negative logratios), and negative values are not allowed for the
  lancaster method.

  This method works around this problem by specifying a weighting function that is compatible with the logratios
  and with the lancaster method. The default is to specify the string 'best', which uses an internal function to
  determine how to exponentiate the logratios to get the ratios, using whatever base was used for the transformation.
}
