% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alr.R
\name{alr_transformation}
\alias{alr_transformation}
\title{Calculate ALR}
\usage{
alr_transformation(mat, denom_name = NULL, delta = NULL,
  impute_proportion = 0.65)
}
\arguments{
\item{mat}{an D x M matrix of D target IDs and
M samples}

\item{denom_name}{a character vector of target IDs or
an integer vector of row numbers. If there is
more than one, the geometric mean of the
denominator values within one sample is used
as the denominator.}

\item{delta}{a number that is the imputed value. If \code{NULL},
delta = impute_proportion * (minimum value in sample)}

\item{impute_proportion}{percentage of minimum value that
becomes the imputed value. Only used if delta is \code{NULL}}
}
\value{
(D - n - z) x M matrix of ALR-transformed
  values, with n equal to the number of denominator values
  and z are the number of rows with essential zeros.
}
\description{
This function calculates the additive logratio
transformation. This function uses calculate_alr
internally after processing the data to handle
zeros. Essential zeros (i.e. target IDs with 
zeros in all samples) are excluded, and
round zeros (i.e. target IDs with at least one
zero value) are imputed using the mulitplicative
method.
}
