% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr.R
\name{clr_transformation}
\alias{clr_transformation}
\title{Centered Logratio Transformation}
\usage{
clr_transformation(mat, base = "e", remove_zeros = FALSE,
  method = "multiplicative", delta = NULL, impute_proportion = 0.65)
}
\arguments{
\item{mat}{N x M matrix of estimated abundances}

\item{base}{what should the base of the logarithm be?
currently only supports base "e" and base 2.}

\item{remove_zeros}{boolean to see if this function
should remove essential zeros (features with zeros in
all samples). The default is \code{FALSE} to be
compatible with sleuth, as its default filter removes
essential zeros.}

\item{method}{which method to use for imputing zeros.
'multiplicative' (default) sets all values smaller than
a imputation value 'delta' (determined by delta or
impute_proportion) to that imputation value, and reduces
all other values by the amount X * (1 - delta*num_zero_values /
sum_constraint). 'additive' is similar to most other tools, and
just adds the imputation value to all entries ('delta' must
be specified)}

\item{delta}{a number that is the imputed value. If \code{NULL},
delta = impute_proportion * (minimum value in sample)}

\item{impute_proportion}{percentage of minimum value that
becomes the imputed value. Only used if delta is \code{NULL}}
}
\value{
N x M matrix of CLR-transformed values with 
essential zero rows removed.
}
\description{
This function applies the centered
logratio transformation on a matrix of
expression values.
}
\details{
this converts an N x M matrix of
N target IDs and M samples (N >> M). 
If M > N, then the matrix is flipped to do 
the calculations, but returned with the same           
as the input. The calculation is as follows:
x_1, x_2, ..., x_D => log(x_1 / g(X)), ..., log(x_D / g(X))
}
