% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr.R
\name{clr_transformation}
\alias{clr_transformation}
\title{Centered Logratio Transformation}
\usage{
clr_transformation(mat, base = "e", remove_zeros = FALSE,
  delta = NULL, impute_proportion = 0.65)
}
\arguments{
\item{mat}{N x M matrix of estimated abundances}

\item{base}{what should the base of the logarithm be?
currently only supports base "e" and base 2.}

\item{delta}{a number that is the imputed value. If \code{NULL},
delta = impute_proportion * (minimum value in sample)}

\item{impute_proportion}{percentage of minimum value that
becomes the imputed value. Only used if delta is \code{NULL}}
}
\value{
N x M matrix of CLR-transformed values with 
essential zero rows removed.
}
\description{
This function applies the centered
logratio transformation on a matrix of
expression values.
}
\details{
this converts an N x M matrix of
N target IDs and M samples (N >> M). 
If M > N, then the matrix is flipped to do 
the calculations, but returned with the same           
as the input. The calculation is as follows:
x_1, x_2, ..., x_D => log(x_1 / g(X)), ..., log(x_D / g(X))
}
