% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sleuth.R
\name{make_lr_sleuth_object}
\alias{make_lr_sleuth_object}
\title{Make Compositional Sleuth Object}
\usage{
make_lr_sleuth_object(sample_to_covariates,
  full_model = stats::formula("~condition"), target_mapping, beta,
  null_model = stats::formula("~1"), aggregate_column = NULL,
  num_cores = parallel::detectCores() - 2, lr_type = "alr",
  denom_name = NULL, ...)
}
\arguments{
\item{sample_to_covariates, }{the sample_to_covariates matrix for sleuth}

\item{full_model, }{the full model for sleuth}

\item{target_mapping, }{the target mapping data frame for sleuth}

\item{beta, }{the beta you wish to use for the Wald test. If NULL,
the Wald test will be skipped.}

\item{null_model, }{the null model to be the baseline for the LR test.
default is the intercept only model (~1).}

\item{aggregate_column, }{character indicating the column in
\code{target_mapping} to be used for "gene-level" sleuth analysis.}

\item{num_cores, }{the number of cores to be used for sleuth analysis}

\item{lr_type, }{either "alr" or "clr" ("ALR" / "CLR" also accepted),
indicating additive logratio or centered logratio transformation}

\item{denom_name, }{target ID names or index numbers of denominators;
required for the ALR transformation.}

\item{...}{extra options that will be passed on the sleuth.
you can specify here whether \code{read_bootstrap_tpm} and
\code{extra_bootstrap_summary} should be \code{FALSE} (default \code{TRUE}).
you can also specify here 'delta' and 'impute_proportion' for
the transformation functions.}
}
\value{
a sleuth object that has been prepped and fitted using the
  full and null models. It will also run the Wald test (if applicable)
  and the LR test
}
\description{
This is a wrapper function for \code{sleuth_prep} that applies
the compositional data analysis approach. Many of the arguments
are ones that will be input into \code{sleuth_prep}.
}
