% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{norm_identity}
\alias{norm_identity}
\title{Norm Identity}
\usage{
norm_identity(mat)
}
\arguments{
\item{mat}{a numeric matrix, with columns for each
sample, and rows for each feature}
}
\value{
a vector of ones with length equal to the
  number of columns (samples) in the input matrix
}
\description{
Function to return size factors of just 1
for sleuth normalization (i.e. lack of
normalization). Normalization is not
necessary for logratio transformations.
}
